added husky, lint-staged, renamed Hero
