+++ cypress/support/commands.ts
@@ -35,3 +35,4 @@
 //     }
 //   }
 // }
+export { };
