+++ package.json
@@ -47,8 +47,10 @@
     "eslint-config-prettier": "^8.7.0",
     "eslint-import-resolver-typescript": "^3.5.3",
     "eslint-plugin-prettier": "^4.2.1",
+    "husky": "^8.0.3",
     "jest": "^29.5.0",
     "jest-environment-jsdom": "^29.5.0",
+    "lint-staged": "^13.2.0",
     "node-mocks-http": "^1.12.1",
     "postcss": "^8.4.21",
     "prettier": "^2.8.4",
