+++ src/components/Hero/MainBanner.tsx
